requirements.txt
Django==3.2                 
django-filter==22.1                
djangorestframework==3.12.4
